Improvements to allow mutliple objects to share the same dc bool object

+++ DCONFIG_Booleans.py
@@ -20,6 +20,10 @@ def create_collection_name(base):
 class DC_MT_boolean_pie(bpy.types.Menu):
     bl_label = "Booleans"
 
+    @classmethod
+    def poll(self, context):
+        return bpy.context.active_object.type == "MESH"
+
     def draw(self, context):
         layout = self.layout
         pie = layout.menu_pie()
@@ -70,49 +74,52 @@ class DC_OT_boolean_live(bpy.types.Operator):
     insetted: bpy.props.BoolProperty(name='Insetted', default=False)
     bool_operation: bpy.props.StringProperty(name="Boolean Operation")
 
-    def rename_boolean_obj(self, base, boolean_obj):
-        new_name = base.name + "_" + self.bool_operation.lower()
-        if self.cutline:
-            new_name += "_cutline"
-
+    def rename_boolean_obj(self, boolean_obj):
         old_name = DC.full_name(boolean_obj)
+        new_name = "dc_bool_obj"
+
         boolean_obj.name = new_name
+        boolean_obj.data.name = new_name
         DC.trace(2, "Renamed {} to {}", old_name, DC.full_name(boolean_obj))
 
-    def make_inset_name(self, base):
-        return base.name + "_inset"
+    def rename_inset_obj(self, base, inset):
+        new_name = base.name + "_inset"
+        inset.name = new_name
+        inset.data.name = new_name
 
-    def create_cutter(self, base, boolean_obj, boolean_move_list, inset_move_list):
-        self.rename_boolean_obj(base, boolean_obj)
+    def create_bool_obj(self, base, boolean_obj, boolean_move_list, inset_move_list):
+        if not boolean_obj.name.startswith("dc_"):
+            self.rename_boolean_obj(boolean_obj)
 
-        if self.cutline:
-            boolean_obj.modifiers.new('Cutline', "SOLIDIFY").thickness = 0.02
+            if self.cutline:
+                mod = boolean_obj.modifiers.new('Cutline', "SOLIDIFY")
+                mod.thickness = 0.02
 
-        if self.insetted:
-            bpy.context.view_layer.objects.active = boolean_obj
-            base.select_set(state=False)
-            boolean_obj.select_set(state=True)
+            if self.insetted:
+                bpy.context.view_layer.objects.active = boolean_obj
+                base.select_set(state=False)
+                boolean_obj.select_set(state=True)
 
-            bpy.ops.object.duplicate()
-            inset = bpy.context.active_object
-            inset.name = self.make_inset_name(base)
+                bpy.ops.object.duplicate()
+                inset = bpy.context.active_object
+                self.rename_inset_obj(base, inset)
 
-            bpy.ops.object.editmode_toggle()
-            bpy.ops.mesh.select_all(action='SELECT')
-            bpy.context.scene.tool_settings.transform_pivot_point = 'INDIVIDUAL_ORIGINS'
-            bpy.ops.transform.resize(value=(0.92, 0.92, 0.92), constraint_axis=(False, False, False), mirror=False, proportional='DISABLED')
-            bpy.context.scene.tool_settings.transform_pivot_point = 'MEDIAN_POINT'
-            bpy.ops.object.editmode_toggle()
+                bpy.ops.object.editmode_toggle()
+                bpy.ops.mesh.select_all(action='SELECT')
+                bpy.context.scene.tool_settings.transform_pivot_point = 'INDIVIDUAL_ORIGINS'
+                bpy.ops.transform.resize(value=(0.92, 0.92, 0.92), constraint_axis=(False, False, False), mirror=False, proportional='DISABLED')
+                bpy.context.scene.tool_settings.transform_pivot_point = 'MEDIAN_POINT'
+                bpy.ops.object.editmode_toggle()
 
-            bpy.context.view_layer.objects.active = boolean_obj
-            bpy.ops.object.constraint_add(type='COPY_TRANSFORMS')
-            bpy.context.object.constraints["Copy Transforms"].target = inset
-            inset_move_list.append(inset)
+                bpy.context.view_layer.objects.active = boolean_obj
+                bpy.ops.object.constraint_add(type='COPY_TRANSFORMS')
+                bpy.context.object.constraints["Copy Transforms"].target = inset
+                inset_move_list.append(inset)
 
         boolean_obj.display_type = 'WIRE'
         boolean_move_list.append(boolean_obj)
 
-    def create_modifier(self, base, boolean_obj):
+    def create_bool_mod(self, base, boolean_obj):
         DC.trace(2, "Adding boolean modifier to {}", DC.full_name(base))
         mod = base.modifiers.new(boolean_obj.name, "BOOLEAN")
         mod.object = boolean_obj
@@ -156,21 +163,18 @@ class DC_OT_boolean_live(bpy.types.Operator):
     def execute(self, context):
         DC.trace_enter("DC_OT_boolean_live.execute")
 
-        # Grab our main active object and validate it's actually a mesh...
+        # Grab our main active object...
         base = bpy.context.active_object
-        if base.type != "MESH":
-            return DC.trace_exit("DC_OT_boolean_live.execute", result="CANCELLED")
-
         DC.trace(1, "Base: {}", DC.full_name(base))
 
-        # Prepare the boolean geometry and our collections...
+        # Prepare the boolean objects and our collections...
         self.prepare_objects()
         col_orig, col_bool = self.prepare_collections(base)
 
         bpy.ops.object.editmode_toggle()
         selected = bpy.context.selected_objects
 
-        DC.trace(1, "Current selection: {}", selected)
+        DC.trace(1, "Current selection: {}", DC.full_names(selected))
 
         # Perform actual boolean operations (keeping track of the final set of geometry to move)...
         boolean_move_list = []
@@ -178,18 +182,19 @@ class DC_OT_boolean_live(bpy.types.Operator):
         for obj in selected:
             if obj != base:
                 DC.trace(1, "Found boolean obj: {}", DC.full_name(obj))
-                self.create_cutter(base, obj, boolean_move_list, inset_move_list)
-                self.create_modifier(base, obj)
+                self.create_bool_obj(base, obj, boolean_move_list, inset_move_list)
+                self.create_bool_mod(base, obj)
 
         # Place everything in the right collection...
-        DC.trace(1, "Boolean Move list: {}", boolean_move_list)
-        DC.trace(1, "Inset Move list: {}", inset_move_list)
+        DC.trace(1, "Boolean Move list: {}", DC.full_names(boolean_move_list))
+        DC.trace(1, "Inset Move list: {}", DC.full_names(inset_move_list))
         base.select_set(state=False)
         for item in boolean_move_list:
             item.select_set(state=False)
             if item.name not in col_bool.objects:
                 col_bool.objects.link(item)
-            col_orig.objects.unlink(item)
+            if item.name in col_orig.objects:
+                col_orig.objects.unlink(item)
 
         for item in inset_move_list:
             if item.name not in col_orig.objects:
@@ -208,11 +213,8 @@ class DC_OT_toggle_cutters(bpy.types.Operator):
     def execute(self, context):
         DC.trace_enter("DC_OT_toggle_cutters.execute")
 
-        # Grab our main active object and validate it's actually a mesh...
+        # Grab our main active object...
         base = bpy.context.active_object
-        if base.type != "MESH":
-            return DC.trace_exit("DC_OT_toggle_cutters.execute", result="CANCELLED")
-
         DC.trace(1, "Base: {}", DC.full_name(base))
 
         # Toggle viewport visibility on our special boolean collection for this object...
@@ -235,11 +237,8 @@ class DC_OT_boolean_apply(bpy.types.Operator):
         if bpy.context.mode != "OBJECT":
             return DC.trace_exit("DC_OT_boolean_apply.execute", result="CANCELLED")
 
-        # Grab our main active object and validate it's actually a mesh...
+        # Grab our main active object...
         base = bpy.context.active_object
-        if base.type != "MESH":
-            return DC.trace_exit("DC_OT_boolean_apply.execute", result="CANCELLED")
-
         DC.trace(1, "Base: {}", DC.full_name(base))
 
         # We need to apply everything up until the last boolean modifier
@@ -263,6 +262,7 @@ class DC_OT_boolean_apply(bpy.types.Operator):
 
         DC.trace(1, "Processing {} orphaned objects", len(orphaned_objects))
 
+        skip_count = 0
         collection_name = create_collection_name(base)
         for col in bpy.data.collections:
             if col.name != collection_name:
@@ -270,6 +270,7 @@ class DC_OT_boolean_apply(bpy.types.Operator):
                     if orphaned_objects[i].name in col.all_objects:
                         DC.trace(2, "Object {} belongs to another collection called {}. Skipping...", orphaned_objects[i].name, col.name)
                         orphaned_objects.remove(orphaned_objects[i])
+                        skip_count = skip_count + 1
 
         if len(orphaned_objects) > 0:
             DC.trace(1, "Removing {} orphaned objects", len(orphaned_objects))
@@ -281,7 +282,7 @@ class DC_OT_boolean_apply(bpy.types.Operator):
 
         if collection_name in bpy.data.collections:
             col = bpy.data.collections[collection_name]
-            if len(col.all_objects) == 0:
+            if len(col.all_objects) - skip_count == 0:
                 DC.trace(1, "Removing collection: {}", collection_name)
                 context.scene.collection.children.unlink(col)
                 bpy.data.collections.remove(col)
